+++ src/Yesod/Auth/Simple.hs
@@ -6,6 +6,7 @@
 {-# LANGUAGE Rank2Types #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE TupleSections #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 
@@ -116,6 +117,9 @@ confirmR = PluginR "simple" ["confirm"]
 confirmationEmailSentR :: AuthRoute
 confirmationEmailSentR = PluginR "simple" ["confirmation-email-sent"]
 
+confirmationEmailResentR :: AuthRoute
+confirmationEmailResentR = PluginR "simple" ["confirmation-email-resent"]
+
 loginR :: AuthRoute
 loginR = PluginR "simple" ["login"]
 
@@ -201,6 +205,13 @@ class (YesodAuth a, PathPiece (AuthSimpleId a)) => YesodAuthSimple a where
   -- if the token is still valid and return the associated email if so.
   matchRegistrationToken :: Text -> AuthHandler a (Maybe Email)
 
+  {- | Сheck if a registration confirmation is pending for the given email.
+    
+    @since 0.1.0
+  -}
+  isConfirmationPending :: Email -> AuthHandler a Bool
+  isConfirmationPending _ = pure False
+
   -- | Like 'matchRegistrationToken' but for password resets.
   matchPasswordToken :: Text -> AuthHandler a (Maybe (AuthSimpleId a))
 
@@ -249,6 +260,13 @@ class (YesodAuth a, PathPiece (AuthSimpleId a)) => YesodAuthSimple a where
 
   confirmationEmailSentTemplate :: WidgetFor a ()
   confirmationEmailSentTemplate = confirmationEmailSentTemplateDef
+
+  {- | Template to notify user that a confirmation email has been resent.
+
+    @since 0.1.0
+  -}
+  confirmationEmailResentTemplate :: WidgetFor a ()
+  confirmationEmailResentTemplate = confirmationEmailSentTemplate
 
   resetPasswordEmailSentTemplate :: WidgetFor a ()
   resetPasswordEmailSentTemplate = resetPasswordEmailSentTemplateDef
@@ -304,6 +322,7 @@ dispatch method path = case (method, path) of
   ("GET",  ["confirm"])                   -> sr getConfirmR
   ("POST", ["confirm"])                   -> sr postConfirmR
   ("GET",  ["confirmation-email-sent"])   -> sr getConfirmationEmailSentR
+  ("GET",  ["confirmation-email-resent"]) -> sr getConfirmationEmailResentR
   ("GET",  ["register-success"])          -> sr getRegisterSuccessR
   ("GET",  ["user-exists"])               -> sr getUserExistsR
   ("GET",  ["login"])                     -> sr getLoginR
@@ -410,7 +429,7 @@ postRegisterR = do
   (honeypot, email) <- runInputPost $ (,)
                       <$> iopt textField honeypotName
                       <*> ireq textField "email"
-  mEmail <- validateAndNormalizeEmail email
+  mEmail <- fmap Email <$> validateAndNormalizeEmail email
   case mEmail of
     _ | isJust honeypot -> do
           onBotPost
@@ -419,13 +438,21 @@ postRegisterR = do
                     \ if the problem persists."
           redirectWithError registerR msg
     Just email' -> do
-      tp <- getRouteToParent
-      renderUrl <- getUrlRender
-      rawToken <- liftIO genToken
-      let url = renderUrl . tp . confirmTokenR $ encodeToken rawToken
-          hashed = hashAndEncodeToken rawToken
-      sendVerifyEmail (Email email') url hashed
-      redirect $ tp confirmationEmailSentR
+      getUserId email' >>= \case
+        -- User with that email already exists
+        Just _  -> do
+          let msg = "This email address is already in use. Please login to your existing account."
+          redirectWithError registerR msg
+        Nothing -> do
+          tp <- getRouteToParent
+          renderUrl <- getUrlRender
+          rawToken <- liftIO genToken
+          let url = renderUrl . tp . confirmTokenR $ encodeToken rawToken
+              hashed = hashAndEncodeToken rawToken
+          route <- bool confirmationEmailSentR confirmationEmailResentR
+                    <$> isConfirmationPending email'
+          sendVerifyEmail email' url hashed
+          redirect $ tp route
     Nothing -> do
       setError "Invalid email address"
       tp <- getRouteToParent
@@ -553,6 +580,15 @@ getConfirmationEmailSentR :: YesodAuthSimple a => AuthHandler a TypedContent
 getConfirmationEmailSentR = selectRep . provideRep . authLayout $ do
   setTitle "Confirmation email sent"
   confirmationEmailSentTemplate
+
+{- | Confirmation to show after resending verification email.
+
+  @since 0.1.0
+-}
+getConfirmationEmailResentR :: YesodAuthSimple a => AuthHandler a TypedContent
+getConfirmationEmailResentR = selectRep . provideRep . authLayout $ do
+  setTitle "Confirmation email resent"
+  confirmationEmailResentTemplate
 
 -- | Confirmation to show after sending password reset email
 getResetPasswordEmailSentR :: YesodAuthSimple a => AuthHandler a TypedContent